split server implementation into separate file #12
